+++ packages/components/CHANGELOG.md
@@ -9,6 +9,7 @@
 ### New Feature
 
 - Added a new `HorizontalRule` component.
+- Added a new `Snackbar` component.
 
 ### Bug Fix